feat(substate)!: widen the substate version to u64
A substate version counts one write per transaction that touches it, so its
ceiling is a throughput ceiling on a single contended substate rather than a
calendar one. u32 exhausts in ~50 days at 1,000 writes/s on one substate and
~1.4 years at 100/s - both rates a hot vault or AMM pool can reach - where u64
puts the ceiling out of reach entirely.

Overflow was also unsafe rather than a clean reject: every increment was a plain
`+ 1`, and the version is the low bytes of `SubstateAddress`, so a wrapped
version collides with the v0 address and reuses a JMT key for a substate that is
already down.

Consensus-breaking. The version is chained into `hash_substate` through Borsh,
which is fixed-width, so every substate value hash changes; `SubstateAddress`
grows 36 -> 40 bytes, so every JMT key changes with it; and the version inside a
transaction's input requirements is part of the signed preimage. Networks must
reset rather than upgrade in place.

Version routing is untouched: the version is a pure suffix of the address and
`to_shard` reads only its first byte.

Also:
- Pin the version fields to a TS `number` so the generated bindings are
  unchanged; ts-rs maps a bare u64 to `bigint`, and a version stays far below
  2^53.
- Widen the SQLite version columns in the wallet and indexer stores. The wallet's
  `vaults` is a foreign key target, so its column is replaced in place rather
  than through a table rebuild, which inside a migration's transaction would
  repoint every referencing table at the temporary name.
- Raise the `ProposedBlockChangeSet` memory guardrail to 15MB; the wider version
  fields in `SubstateChange` and `SubstateLock` move the worst case from
  13.4MB to 14.3MB.

+++ clients/validator_node_client/src/types.rs
@@ -417,7 +417,8 @@ pub struct GetStateResponse {
 )]
 pub struct GetSubstateRequest {
     pub address: SubstateId,
-    pub version: u32,
+    #[cfg_attr(feature = "ts", ts(type = "number"))]
+    pub version: u64,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
